WiP [g5k] allow "non deploy" reservation
`job_type` can be passed in the configuration. Its value can
be any job type supported but this mainly should be `deploy`
or `allow_classic_ssh`.
The former correspond to the previous situation.
The latter is new. This will reserve production node, this
won't call kadeploy3 but instead user will be granted root
access to the nodes.  Starting from this point, user can
interact as usual with its nodes.