Introduce drivers to interact with G5K
A Driver is reponsible for reserving, reloading or destroying the
resources. It can interact for example with OAR, OARGRID or the REST
API. Currently two drivers that interact with OARGRID are implemented:

- a dynamic (OargridDynamicDriver) that reserves resources according the
the provider configuration

- a static (OargridStaticDriver) that only reloads the resources given
an oargrid jobid. The reservation must be handled externally.

Drivers are implemented in the driver.py file.
The function ``get_driver` reads the provider configuration and load the
relevant driver.

For #49