Relax a bit the schema of Chameleon
It's shared between chameleonkvm (ckvm) and chameleonbaremetal (cbm). cbm needs
some more keys in comparison to ckvm. Logically speaking cbm inheritate from
ckvm and instead of duplicating the whole schema, we choose to relax a bit so
that this inheritance can occur also at the schema level. (It seems that there
isn't an obvious mechanism to map the inheritance from the objet layer to the
schema level).