[vagrant/vmong5k] Favour flavour_desc for the size reference
We ensure that `flavour_desc` dict is set by the MachineConfiguration.
This can be used in the provider to access the flavour details without relying
on the flavour name.

We also unify the key naming. (cpu -> core in both cases).

Not address here: code duplication between the two may be addressed in the
future.