[vmong5k] Allow to choose a working dir
This working dir is used to generate all the files and disks needed by the
provider. It must be an absolute path.

This allows the user to choose whether to store the disks on the local
filesystem on the node (working_dir=/tmp/enos_vmong5k -- this is the default)
or in the site frontend (e.g. working_dir=/home/msimonin/public/enos_vmong5k).

In addition, the vms are now launched as a regular user (the current g5k user).

Be aware that the permissions on the working dir are set to 711 whereas the
permission on the actual vm disks are set to 777 (inside the working
directory).

Fix #5