[VMonG5K] Add strategy keyword
This gouverns how the disk image will be created.
Currently we support:

- copy: `image` will be copied in the `working_dir`. Then a copy
will be made for every single vm that need to boot on the current host.
(that makes n+1 copy of the base image)

- cow: `image` will be copied in the `working_dir`. Then a copy on write
file is created for every single vm that need to boot on the current
host.