[g5k/vmong5k] Support for slash_16 subnets
In the resource data structure, we now allow a concrete network (_c_network) to
be actually a list of networks. Indeed on Grid'5000 a slash_16 network is made
of 64 slash_22 networks. By doing this we can support slash_16 networks. There
might be other way of doing it like creating a true ConcreteNetwork
corresponding to the /16 subnet and hide there the list.