api.play_on: Clarify the use of `gather_facts`
Situation before this patch.

The `gather_facts` option was supposed to control whether the facts are
gathered prior to the execution of the modules.
It could take two values:
- False -> facts was not gathered at all
- True -> facts for the hosts maching the provided `pattern_hosts` were
gathered. Note that this contradicted the docstring who stated that the
facts were supposed to be gathered on all hosts

Situation with this patch.

`gather_fatcs` is now an Union of boolean and string:
- True -> for the hosts maching the provided `pattern_hosts` were
gathered
- False -> no fact will be gatherd
- pattern_hosts: str -> facts will be gathered for hosts that match `pattern`.