svc/netem: support for bridged network
Rationale:
On Grid'5000 default production environment is configured with a bridged
network (for virtualisation purpose). But, tc rules can't be applied on
software bridge (to the best of my knowledge).

Implementation:
Ansible network fact contains the information of bridged interface. E.g. if br0
is a bridged over ~eth1~ the ~br0~ ansible network fact will have an
`interfaces` key equals to ~[eth1]~. We leverage this to pass the correct
physical interface to the ansible  playbook enforcing the network constraints.