VMonG5k: support for extra disk
At low level one can use the `extra_device` parameter key of the configuration
of a machine and write here the libvirt xml corresponding the the wanted
device.
At higher level one can specify a disk parameter in the `flavour_desc`
parameter key.

Note that in the current implementation, the disks are not formatted nor
mounted by EnOSlib.

For #58