service: Don't change default python in service
Before we had a side-effect on some service that (force) sets the default
python interpreter. Since Ansible 2.8 the python interpreter can be passed as a
variable. So we set this on demand for services that require python3 by using
the system provided python3 (but not changing it).
As python3 should be now the default, we'll probably remove any reference to
python3 in the future.