Enostask: rewrite + tests
This is a rewrite of the task module. This module contains the logic that
implement the concepts of `enostask` and `environment`.
An enostask is an experimental facility that allows to keep track of the
experiment context easily. Environment is saved (resp. restored) at the end
(resp. the start) of a task transparently.
This has been around since the beginning but needed some updates. The main
drawback with the initial code was that nested `enostask` wasn't supported.
Additionnaly the code was a bit hard to follow and not tested.

This patch is a rewrite that
- Introduces the environment object (in the past this was a dict)
- Increases the test coverage
- Allows for nested tasks

From the application code this change *should* be transparent. Nevertheless we
change the dump format of the environment to `pickle` (`yaml` was used before).