Netem: support command chunking
For large deployment the command to run on the remote hosts can become very
long and exceed the OS limit. We provide a way to chunk this command using the
`chunk_size` kwargs of the Netem service. The default value (100) seems really
conservative. Increasing this value will decrease the deployment time (with the
risk to hit the limit).

In the future, constant time deployment may be envision by first uploading a script
with all the commands and then running it.