Fix issue with machine's extra_devices property in VMonG5K config
In VMonG5K provider, extra_device is an optionnal property. A bug in the configuration management caused the deployment to raise an exception if extra_device was not set. Indeed, the property was still used even if not set (causing trying to validate that a None attribute was a string)