service/conda: introduce new Dask deployment.
This will manage the dask agents life cycle.
But this assumes conda to be set up in the resources
prior to the deployment. This obviously speed up the deployment.

In G5k this can be achieved by setting up your conda environment in a frontend
(shared space) and running the Dask service on the production nodes
(allow_classic_ssh job type).

The service follows the standard (deploy, destroy) service definition, but it
can also be used as a context manager.

Also:

- G5k: introduce G5kTunnel
- make g5k_api_utils available in __all__