Allow custom stdout callback to be used.
We used to inject a dedicated stdout callback when running a play.
It seems to prevent custom stdout callback to be used (in the config file for
instance).

Now we inject a custom plugin in the list of available plugins instead. This
allows the user the use a custom stdout callback.

Implementation-wise: this is a bit hacky since we access the internal state of
the Ansible's TaskQeueManager