api: use CallbackBase as base class
We inherited from the wrong parent class.
Before, CallbackModule was used but displayed the output of the play execution
on the screen (whatever the Ansible configuration)
Now, we inherit from the (true) base class which does nothing.

Side effect: stdout_callback configuration will work as expected.
(it used to work but the effect was hindered by the wrong base class used)