g5k: new _test_slot breaks the multisite support
The cluster(s) status now needs to be injected in the parameters
Factorize the corresponding code as it is used in `provider.test_slot` and in the
`g5k_api_utils._do_synchronize`.

Status has returne by python-grid5000 and used throughout this code has a
`nodes` attributes. Adapting the code accordingly (unittest and _test_slot).