g5k: fix check_deployed_nodes condition
For some reason, we get SSH access to the "deployment mini OS" while
kadeploy is running.  The current check_deployed condition was wrongly
believing in that case that the host is already deployed, which is not the
case.

This bug can be triggered by interrupting Enoslib while a deployment is
running, and then starting Enoslib again on the same job.

While at it, also change the mount regexp so that it works for NVMe drives
(some recent Grid'5000 clusters have NVMe drives as first disk)