g5k: Fix deployment logging
The three deployment retries are logged as 0/3, 1/3, and finally 2/3.  Fix
this by starting to count at 1/3.

Also, _deploy() is recursive and its last call is always with an empty
list of machines to deploy.  In that case we would get a weird log output:

    Deploying [] with options ...

Fix logging to remove this last message.