Modernize and simplify `ensure_python3`
Since d9894854b7df2ad5f32195d7797b555e27f6c880 (Enoslib 7.0.0), we
already explicitly set the python interpreter to python3.

As a result, we don't need to check that a "python" binary actually
exists, "python3" is enough.  As a matter of fact, modern distros ship
with a "python3" binary but not "python".

Furthermore, we only need python3 itself, not python3-pip (which pulls in
a huge amount of packages through build-essential and python3-dev)

Finally, stop creating a python3 -> python symlink by default, since we
don't need it in Enoslib.