Fix Ansible dependency and allow 7.1 versions
Using a constraint such as "ansible<=7" is a bad idea, because it will
limit the version to exactly 7.0.0 and not even allow 7.0.1.

While fixing this, update to ansible 7.1 as it has only minor core changes.

In the future, we could update to "ansible<8" because 7.(X+1) version
updates of Ansible should remain backwards compatible, but in the interest
of reproducibility, let's be very conservative.

Ref: https://docs.ansible.com/ansible/devel/reference_appendices/release_and_maintenance.html#id13