g5k: make sure check_deployment outputs nothing
Enoslib users can select a non-default output style for Ansible:
`set_config(ansible_stdout="XXX")`

In that case, especially with the default Ansible style, check_deployment
outputs a big scary error:

    TASK [Check deployment] ******************************
    fatal: [gros-8-kavlan-4.nancy.grid5000.fr]: UNREACHABLE! => {"changed": false, "msg": "Failed to connect to the host via ssh: Connection timed out during banner exchange\r\nConnection to UNKNOWN port 65535 timed out", "unreachable": true}

But failures in this task are expected and handled correctly by Enoslib.
To avoid scaring the user for no good reason, we now completely suppress
the output of this task.