Planning Service
The planning service allow to schedule events on different machines.
This is process based. In the future we might want to control different
encapsulations of a program (containers like in
[MockFog2](https://doi.org/10.1109/TCC.2021.3074988).

To make it work we use cron and cgroups.
An addition has been made to the `api.run_command` function that allows
a process to be started within a cgroup. We also added some utils
functions to manage cgroups.

2 tutorials have been added to the tutorial repo.

Co-authored-by: wtrabels <weissem.trabelsi@inria.fr>