g5k: Fix race condition when destroying and reloading jobs
Destroying a job is asynchronous on Grid'5000, so trying to destroy and
recreate a job often ends up reloading the same job that is being
terminated, failing later with an exception:

    Exception: The job XXX is in error state

We should wait for the job to be killed before returning from destroy().

Also add CI test cases for this issue.