vmong5k: remove gateway parameter, use the global g5k_auto_jump setting
The new setting is more flexible, because it allows to force-enable or
force-disable the automatic SSH jump host configuration.  In contrast, the
"gateway" parameter was always enabled when outside G5K.

In addition, it's simpler to have a single setting that applies to both
the G5K and VMonG5K providers.

This will break compatibility for experiment code that was using this
"gateway" parameter.