VMonG5K: Consider memory when allocating VMs on physical hosts
Up to now, we were only considering CPU cores (or threads) to determine
how many physical hosts are required to host the VMs.

We now also consider the amount of memory.  We take a safety margin of 2%
(down to 256 MiB for small nodes and up to 4 GiB for big nodes) to leave
some memory free for the host system.

Fixes: #180