VMonG5K: Fix VM allocation algorithm for very large VM sizes
When asking for VMs that are bigger than the physical hosts, we currently
return more physical hosts than the number of VMs.  This makes no sense,
because a VM cannot be split across multiple physical hosts.

Instead, return as many physical hosts as the number of VMs.  Running VMs
larger than the physical host may or may not work in practice, but that's
not the problem of the allocator.