Docker: don't use sudo to install nvidia-toolkit
In Enoslib, we already assume we are root, there is no need to use sudo.
It wouldn't work anyway as non-root because we were calling "apt" without sudo.

This fixes Enoslib when running minimal images such as debian11-min on
Grid'5000, because sudo is not installed by default, so this crashes the
docker-nvidia installation process.