g5k: allow any job types instead of restricting to a known list
New job types are added regularly, such as Kwollect-specific job types to
select metrics to monitor, or "origin=" to track which tool has created
the job.

Let's remove all constraints on the job type and trust the G5K API
instead: it returns a rather clear error whenever a job type is not
supported.  Example error:

    enoslib.errors.InvalidReservationCritical: 400: Request to https://api.grid5000.fr/stable/sites/nantes/internal/oarapi/jobs.json failed with status 400: {
       "title" : "Bad query",
       "code" : 400,
       "message" : "Error: unknown job type: deplooooy\n\n"
    }