g5k: fix ssh connection timing out after deployement in kavlan
After a deployement, the function dhcp_networks() tries to connect to nodes to use dhclient cmd on secondary interfaces. When the primary interface is deployed in a kavlan, it can take some time to be reachable. This cause the ssh connection to timeout when running the dhcp_networks() function, which makes the program crash. By using the wait_for() function, we make sure every node is reachable before trying to connect to them